Added new roles, permissions and politics to Post

+++ app/Http/Controllers/Auth/AuthController.php
@@ -59,7 +59,13 @@ class AuthController extends Controller
 
     public function logout(Request $request): JsonResponse
     {
-        $request->user()->currentAccessToken()->delete();
+        if (
+            $request->user()
+            &&
+            $request->user()->currentAccessToken()
+        ) {
+            $request->user()->currentAccessToken()->delete();
+        }
 
         return response()->json([
             'message' => 'Logged out',
